Moved Header components to own folder
