+++ src/Screens/Setting/SettingScreen.js
@@ -1,6 +1,6 @@
 import { View, Text } from 'react-native'
 import React from 'react'
-import Breadcrumb from '../../Shared/Components/Breadcrumb'
+import Breadcrumb from '../../Shared/Components/header/Breadcrumb'
 
 const SettingScreen = () => {
   return (
